Fixed Real-time vaultlist toggling
