+++ apps/app/README.md
@@ -31,6 +31,7 @@ Open [http://localhost:3000](http://localhost:3000) on your browser to see the r
 - `components` - React components that make up the contents of the pages.
 - `hooks` - App-specific hooks not included in other hook packages.
 - `constants` - Constant values, references and configurations to deploy and run this app.
+- `vaultLists` - Any default vault lists for the app.
 
 ## App Data Configuration
